API documentation

Structures

CitableParserBuilder.AbbreviatedUrn — Type

Short form of a Cite2Urn containing only collection and object ID.

CitableParserBuilder.Stem — Type

Supertype of all concrete Stem structures.

CitableParserBuilder.Rule — Type

Supertype of all concrete Rule structures.

CitableParserBuilder.Analysis — Type

Citable analysis of a string value.

An Analysis has five members: a token string value, and four abbreviated URNs, one each for the lexeme, form, rule and stem.

CitableParserBuilder.StemUrn — Type

Abbreviated URN for a morphological stem.

CitableParserBuilder.RuleUrn — Type

Abbreviated URN for rule.

CitableParserBuilder.LexemeUrn — Type

Abbreviated URN for a lexeme.

CitableParserBuilder.FormUrn — Type

Abbreviated URN for a morphological form.

CitableParserBuilder.AnalyzedToken — Type

Morphological analyses for a token identified by CTS URN.

Parsing

CitableParserBuilder.parsetoken — Function

Delegate to specific functions based on type's citable trait value.

parsetoken(s, x; data)

It is an error to invoke the parsetoken using types that are not a parser.

parsetoken(, s, x; data)

Citable parsers must implement parsetoken.

parsetoken(, s, x; data)

Parse String s by looking it up in a given dictionary.

CitableParserBuilder.parsewordlist — Function

Parse a list of tokens with a CitableParser.

parsewordlist(vocablist, p; data, countinterval)

Returns a Dict mapping strings to a (possibly empty) vector of Analysis objects.

CitableParserBuilder.parselistfromfile — Function

Parse a list of tokens in a file with a CitableParser.

parselistfromfile(f, p)
parselistfromfile(f, p, delim; data)

CitableParserBuilder.parselistfromurl — Function

Parse a list of tokens at a given url with a CitableParser.

CitableParserBuilder.parsepassage — Function

Parse a CitablePassage with text for a single token with a CitableParser.

parsepassage(cn, p; data)

Should return an AnalyzedToken.

CitableParserBuilder.parsecorpus — Function

Use a CitableParser to parse a CitableTextCorpus with each citable node containing containg a single token.

parsecorpus(c, p; data, countinterval)

Should return a list of AnalyzedTokens.

Profiling parsed results

CitableParserBuilder.vocabulary_density — Function

Ratio of distinct tokens to size of corpus.

vocabulary_density(tp)

Compute ratio of distinct tokens to size of corpus.

vocabulary_density(tc)

CitableParserBuilder.token_coverage — Function

Percentage of all tokens parsed.

token_coverage(tp)

Compute percentage of all tokens parsed.

token_coverage(tc)

CitableParserBuilder.vocabulary_coverage — Function

Percentage of unique tokens parsed.

vocabulary_coverage(tp)

Compute percentage unique tokens parsed.

vocabulary_coverage(tc)

CitableParserBuilder.lexical_density — Function

Ratio of lexicon size to number of tokens parsed.

lexical_density(tp)

Compute ratio of lexicon size to number of tokens parsed.

lexical_density(tc)

CitableParserBuilder.form_density_incorpus — Function

Ratio of tokens to distinct forms.

form_density_incorpus(tp)

Compute ratio of tokens to distinct forms.

form_density_incorpus(tc)

CitableParserBuilder.form_density_invocabulary — Function

Ratio of vocabulary items to distinct forms.

form_density_invocabulary(tp)

Compute ratio of vocabulary items to distinct forms.

form_density_invocabulary(tc)

CitableParserBuilder.form_density_inlexicon — Function

Ratio of lexicon to distinct forms.

form_density_inlexicon(tp)

Compute ratio of lexicon to distinct forms.

form_density_inlexicon(tc)

CitableParserBuilder.formal_ambiguity — Function

Ratio of distinct forms to all tokens parsed.

formal_ambiguity(tp)

Compute ratio of distinct forms to all tokens parsed.

formal_ambiguity(tc)

Compute morphological ambiguity in corpus analyzed by parser.

formal_ambiguity(c, p; data)

Compute morphological ambiguity in list of words analyzed by parser.

formal_ambiguity(vocablist, p; data)

CitableParserBuilder.lexical_ambiguity — Function

Ratio of lexically ambiguous tokens to all parsed tokens.

lexical_ambiguity(tp)

Compute ratio of lexically ambiguous tokens to all parsed tokens.

lexical_ambiguity(tc)

Compute lexical ambiguity in corpus analyzed by parser.

lexical_ambiguity(c, p; data)

Compute lexical ambiguity in list of words analyzed by parser.

lexical_ambiguity(vocablist, p; data)

Working with vectors of AnalyzedTokens

CitableParserBuilder.lexemes — Function

Extract a list of lexeme values from a Vector of AnalyzedTokens.

lexemes(v)

CitableParserBuilder.stringsforlexeme — Function

Find token string values for all tokens in a vector of AnalyzedTokens parsed to a given lexeme.

stringsforlexeme(v, l)

CitableParserBuilder.lexemedictionary — Function

From a vector AnalyzedTokens and an index of tokens in a corpus, construct a dictionary keyed by lexemes, mapping to a further dictionary of surface forms to passages.

lexemedictionary(parses, tokenindex)

Working with AbbreviatedUrns

CitableParserBuilder.abbreviate — Function

Constructs an AbbreviatedUrn string from a Cite2Urn.

abbreviate(urn)

Example:

julia> abbreviate(Cite2Urn("urn:cite2:kanones:lsj.v1:n123"))
"lsj.n123"

Example: a pipeline abbreviating a Cite2Urn and forming a LexemeUrn from the abbreviated string value.

julia> Cite2Urn("urn:cite2:kanones:lsj.v1:n123") |> abbreviate |> LexemeUrn
LexemeUrn("lsj", "n123")

CitableParserBuilder.expand — Function

Constructs a Cite2Urn from an AbbreviatedUrn and a dictionary mapping collection identifiers in AbbreviatedUrns's to full Cite2Urns for a versioned collection.

CitableParserBuilder.fstsafe — Function

Compose SFST representation of an AbbreviatedUrn.

fstsafe(au)

Example:

julia> LexemeUrn("lexicon.lex123") |> fstsafe
"<u>lexicon\.lex123</u>"

Working with Stems and Rules

CitableParserBuilder.lexeme — Function

Function required to get lexeme value of a Stem implementation.

CitableParserBuilder.id — Function

Function required to get ID value of a Stem implementation.

Function required to get ID value of a Rule implementation.

CitableParserBuilder.inflectiontype — Function

Function required to get string value for inflection class of a Stem implementation.

Function required to get string value for inflection class of a Rule implementation.

Serialization

CitableParserBuilder.readfst — Function

Read SFST output from file f, and parse into a dictionary keying tokens to a (possibly empty) array of SFST strings.

readfst(f)

CitableParserBuilder.analyses_relationsblock — Function

Compose a CEX relationset block for a set of analyses.

analyses_relationsblock(urn, label, v)
analyses_relationsblock(urn, label, v, delim; registry)

CitableParserBuilder.delimited — Function

Serialize an Analysis to delimited text. Abbreviated URNs are expanded to full CITE2 URNs using registry as the expansion dictionary.

delimited(a)
delimited(a, delim; registry)

Serialize a Vector of Analysis objects as delimited text.

delimited(v)
delimited(v, delim; registry)

Serialize a single AnalyzedToken as one or more lines of delimited text.

delimited(at)
delimited(at, delim; registry)

Serialize a Vector of AnalyzedTokens as delimited text.

delimited(v)
delimited(v, delim; registry)

CitableBase.cex — Function

Delegate to specific functions based on type's citable trait value.

cex(x)
cex(x, delim)

It is an error to invoke the cex function on material that is not citable.

cex(_, x, delim)

Citable text content should implement cex.

cex(_, txt, delim)

Citable content should implement cex.

cex(_, obj, delim)

Format a CitablePassage as a delimited-text string.

cex(psg)
cex(psg, delim)

Required function for Citable abstraction.

Format a CitableDocument as a CEX ctsdata block.

cex(doc)
cex(doc, delim)

Required function for Citable abstraction.

Compose a delimited-text string for a corpus.

cex(c; delimiter)

Implement cex function required by Citable interface for CatalogedText.

cex(cataloged)
cex(cataloged, delimiter)

Serialize an AnalyzedToken as delimited text (required for Citable interface).

cex(at)
cex(at, delim)

Uses abbreviated URNs. These can be expanded to full CITE2 URNs when read back with a URN registry, or the delimited function can be used with a URN registry to write full CITE2 URNs.