API documentation

Structures

CitableParserBuilder.AbbreviatedUrn — Type

Short form of a Cite2Urn containing only collection and object ID.

CitableParserBuilder.Stem — Type

Supertype of all concrete Stem structures.

CitableParserBuilder.Rule — Type

Supertype of all concrete Rule structures.

CitableParserBuilder.Analysis — Type

A citable morphological analysis.

An Analysis has five members: a token string value, and four abbreviated URNs, one each for the lexeme, form, rule and stem.

CitableParserBuilder.StemUrn — Type

Abbreviated URN for a morphological stem.

CitableParserBuilder.RuleUrn — Type

Abbreviated URN for rule.

CitableParserBuilder.LexemeUrn — Type

Abbreviated URN for a lexeme.

CitableParserBuilder.FormUrn — Type

Abbreviated URN for a morphological form.

CitableParserBuilder.AnalyzedToken — Type

Morphological analyses for a token identified by CTS URN.

Parsing

CitableParserBuilder.parsetoken — Function

Required function to parse a single token with a CitableParser.

parsetoken(p, t)

Should return a (possibly empty) Vector of Analyses.

CitableParserBuilder.parsewordlist — Function

Parse a list of tokens with a CitableParser.

parsewordlist(p, tokens)

Should return a (possibly empty) Vector of Analyses.

CitableParserBuilder.parselistfromfile — Function

Parse a list of tokens in a file with a CitableParser.

Should return pairings of tokens with a (possibly empty) Vector of Analyses.

CitableParserBuilder.parselistfromurl — Function

Parse a list of tokens at a given url with a CitableParser.

Should return pairings of tokens with a (possibly empty) Vector of Analyses.

CitableParserBuilder.parsenode — Function

Parse a CitableNode with text for a single token with a CitableParser.

parsenode(p, cn)

Should return a pairing of the CitableNode with a list of analyses.

CitableParserBuilder.parsecorpus — Function

Use a CitableParser to parse a CitableTextCorpus with each citable node containing containg a single token.

parsecorpus(p, c)

Should return a list of pairings of a CitableNode with a list of analyses.

Creating an analyzed corpus

CitableParserBuilder.analyzecorpus — Function

Pair each lexical node in a token-level edition with a list of Analysis objects for that token.

analyzecorpus(parser, ortho, c)

Working with AbbreviatedUrns

CitableParserBuilder.abbreviate — Function

abbreviate(urn::Cite2Urn)

Constructs an AbbreviatedUrn string from a Cite2Urn.

Example:

julia> abbreviate(Cite2Urn("urn:cite2:kanones:lsj.v1:n123"))
"lsj.n123"

Example: a pipeline abbreviating a Cite2Urn and forming a LexemeUrn from the abbreviated string value.

julia> Cite2Urn("urn:cite2:kanones:lsj.v1:n123") |> abbreviate |> LexemeUrn
LexemeUrn("lsj", "n123")

CitableParserBuilder.expand — Function

expand(au::AbbreviatedUrn, registry::Dict)

Constructs a Cite2Urn from an AbbreviatedUrn and a dictionary mapping collection identifiers in AbbreviatedUrns's to full Cite2Urns for a versioned collection.

CitableParserBuilder.fstsafe — Function

Compose SFST representation of an AbbreviatedUrn.

fstsafe(au)

Example:

julia> LexemeUrn("lexicon.lex123") |> fstsafe
"<u>lexicon\.lex123</u>"

CitableParserBuilder.abbreviation — Function

Create string abbreviation for an AbbreviatedUrn.

abbreviation(au)

Working with Stems and Rules

CitableParserBuilder.lexeme — Function

Function required to get lexeme value of a Stem implementation.

CitableParserBuilder.id — Function

Function required to get ID value of a Stem implementation.

Function required to get ID value of a Rule implementation.

CitableParserBuilder.inflectiontype — Function

Function required to get string value for inflection class of a Stem implementation.

Function required to get string value for inflection class of a Rule implementation.

Reading FST output from a file

CitableParserBuilder.readfst — Function

Read SFST output from file, and parse into a dictionary of tokens -> (possibly empty) array of SFST strings.

readfst(f)

Parameters

f Name of file to read in.

Serialization

CitableParserBuilder.cex — Function

Serialize an Analaysis as delimted text.

cex(a)
cex(a, delim)

Serialize an Analaysis as delimited text.

cex(tkn; delim, delim2, delim3)

Compose delimited text representation for a map of tokens to a vector of analyses.

cex(prs)