API documentation

Functions

CitableParserBuilder.abbreviate — Method

abbreviate(urn::Cite2Urn)

Constructs an AbbreviatedUrn string from a Cite2Urn.

Example:

julia> abbreviate(Cite2Urn("urn:cite2:kanones:lsj.v1:n123"))
"lsj.n123"

Example: a pipeline abbreviating a Cite2Urn and forming a LexemeUrn from the abbreviated string value.

julia> Cite2Urn("urn:cite2:kanones:lsj.v1:n123") |> abbreviate |> LexemeUrn
LexemeUrn("lsj", "n123")

CitableParserBuilder.expand — Method

expand(au::AbbreviatedUrn, registry::Dict)

Constructs a Cite2Urn from an AbbreviatedUrn and a dictionary mapping collection identifiers in AbbreviatedUrns's to full Cite2Urns for a versioned collection.

CitableParserBuilder.fstsafe — Method

fstsafe(au::AbbreviatedUrn)

Compose SFST representation of an AbbreviatedUrn.

Example:

julia> LexemeUrn("lexicon.lex123") |> fstsafe
"<u>lexicon\.lex123</u>"

Structures

CitableParserBuilder.LexemeUrn — Type

Abbreviated URN for a lexeme.

CitableParserBuilder.StemUrn — Type

Abbreviated URN for a morphological stem.

CitableParserBuilder.RuleUrn — Type

Abbreviated URN for rule.

CitableParserBuilder.FormUrn — Type

Abbreviated URN for a morphological form.

CitableParserBuilder.Analysis — Type

A citable morphological analysis.

An Analysis has five members: a token string value, and four abbreviated URNs, one each for the lexeme, form, rule and stem.